feat(): added nav bar to Layout, modified sign in page, modified cart page
