+++ app/routes/__theme/details/$id.tsx
@@ -154,7 +154,7 @@ export default function Example() {
                     ·
                   </div>
                   <div className="ml-4 flex">
-                    <a href="#" className="text-sm font-medium text-indigo-600 hover:text-indigo-500">
+                    <a href="#" className="text-sm font-medium text-orange-600 hover:text-orange-500">
                       See all {product.reviewCount} reviews
                     </a>
                   </div>
