+++ app/routes/__theme/homepage.tsx
@@ -1,0 +1,176 @@
+import { Dialog, Popover, Tab, Transition } from "@headlessui/react";
+import {
+  MenuIcon,
+  SearchIcon,
+  ShoppingBagIcon,
+  XIcon,
+} from "@heroicons/react/outline";
+import { Link } from "@remix-run/react";
+import serenuplogo from "public/assets/seren-up-logo.png";
+import bracelets from "public/assets/bracelets.jpg";
+import { Fragment, useState } from "react";
+
+const footerNavigation = {
+  shop: [
+    { name: "Bracelets", href: "/products" },
+  ],
+  company: [
+    { name: "Who we are", href: "#" },
+    { name: "Terms & Conditions", href: "#" },
+    { name: "Privacy", href: "#" },
+  ],
+  account: [{ name: "Manage Account", href: "#" }],
+  connect: [{ name: "Contact Us", href: "#" }],
+};
+
+export default function Homepage() {
+  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
+
+  return (
+    <div className="bg-white">
+      {/* Mobile menu */}
+
+
+      {/* Hero section */}
+      <div className="relative bg-orange-700">
+        {/* Decorative image and overlay */}
+        <div aria-hidden="true" className="absolute inset-0 overflow-hidden">
+          <img
+            src={bracelets}
+            alt=""
+            className="w-full h-full object-center object-cover"
+          />
+        </div>
+        <div
+          aria-hidden="true"
+          className="absolute inset-0 bg-gradient-to-b from-orange-600 to-orange-100  opacity-20"
+        />
+
+        {/* Navigation */}
+        <header className="relative z-10">
+
+        </header>
+
+        <div className="relative max-w-3xl mx-auto py-32 px-6 flex flex-col items-center text-center sm:py-64 lg:px-0">
+          <h1 className="text-4xl font-extrabold tracking-tight text-white lg:text-6xl drop-shadow-2xl shadow-zinc-500">
+            New arrivals are here
+          </h1>
+          <p className="mt-4 text-xl bg-orange-400 text-white drop-shadow-xl shadow-black rounded-full px-8 py-1">
+            The new arrivals have, well, newly arrived. <br />
+            Check out the latest options from our products page.
+          </p>
+        </div>
+      </div>
+      <footer aria-labelledby="footer-heading" className="bg-orange-900">
+        <h2 id="footer-heading" className="sr-only">
+          Footer
+        </h2>
+        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+          <div className="py-20 xl:grid xl:grid-cols-3 xl:gap-8">
+            <div className="grid grid-cols-2 gap-8 xl:col-span-2">
+              <div className="space-y-12 md:space-y-0 md:grid md:grid-cols-2 md:gap-8">
+                <div>
+                  <h3 className="text-sm font-medium text-white">Shop</h3>
+                  <ul role="list" className="mt-6 space-y-6">
+                    {footerNavigation.shop.map((item) => (
+                      <li key={item.name} className="text-sm">
+                        <a
+                          href={item.href}
+                          className="text-gray-300 hover:text-white"
+                        >
+                          {item.name}
+                        </a>
+                      </li>
+                    ))}
+                  </ul>
+                </div>
+                <div>
+                  <h3 className="text-sm font-medium text-white">Company</h3>
+                  <ul role="list" className="mt-6 space-y-6">
+                    {footerNavigation.company.map((item) => (
+                      <li key={item.name} className="text-sm">
+                        <a
+                          href={item.href}
+                          className="text-gray-300 hover:text-white"
+                        >
+                          {item.name}
+                        </a>
+                      </li>
+                    ))}
+                  </ul>
+                </div>
+              </div>
+              <div className="space-y-12 md:space-y-0 md:grid md:grid-cols-2 md:gap-8">
+                <div>
+                  <h3 className="text-sm font-medium text-white">Account</h3>
+                  <ul role="list" className="mt-6 space-y-6">
+                    {footerNavigation.account.map((item) => (
+                      <li key={item.name} className="text-sm">
+                        <a
+                          href={item.href}
+                          className="text-gray-300 hover:text-white"
+                        >
+                          {item.name}
+                        </a>
+                      </li>
+                    ))}
+                  </ul>
+                </div>
+                <div>
+                  <h3 className="text-sm font-medium text-white">Connect</h3>
+                  <ul role="list" className="mt-6 space-y-6">
+                    {footerNavigation.connect.map((item) => (
+                      <li key={item.name} className="text-sm">
+                        <a
+                          href={item.href}
+                          className="text-gray-300 hover:text-white"
+                        >
+                          {item.name}
+                        </a>
+                      </li>
+                    ))}
+                  </ul>
+                </div>
+              </div>
+            </div>
+            <div className="mt-12 md:mt-16 xl:mt-0">
+              <h3 className="text-sm font-medium text-white">
+                Sign up for our newsletter
+              </h3>
+              <p className="mt-6 text-sm text-gray-300">
+                The latest deals and savings, sent to your inbox weekly.
+              </p>
+              <form className="mt-2 flex sm:max-w-md">
+                <label htmlFor="email-address" className="sr-only">
+                  Email address
+                </label>
+                <input
+                  id="email-address"
+                  type="text"
+                  autoComplete="email"
+                  required
+                  className="appearance-none min-w-0 w-full bg-white border border-white rounded-md shadow-sm py-2 px-4 text-base text-gray-900 placeholder-gray-500 focus:outline-none focus:border-white focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-900 focus:ring-white"
+                />
+                <div className="ml-4 flex-shrink-0">
+                  <button
+                    type="submit"
+                    className="w-full bg-orange-600 border border-transparent rounded-md shadow-sm py-2 px-4 flex items-center justify-center text-base font-medium text-white hover:bg-orange-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-white focus:ring-orange-500"
+                  >
+                    Sign up
+                  </button>
+                </div>
+              </form>
+            </div>
+          </div>
+
+          {/*references to seren-up and Systax */}
+          <div className="border-t border-gray-800 py-10">
+            <p className="text-sm text-gray-400">
+              Copyright &copy; 2022 Seren-Up S.p.A. - Made By Systax
+            </p>
+          </div>
+        </div>
+      </footer>
+    </div>
+  );
+}
