+++ app/routes/__theme/products.tsx
@@ -123,9 +123,10 @@ type ElaboratedDataProps = {
 export default function Products() {
   const { listableProductList } = useLoaderData<LoaderData>();
   return (
-    <div className="bg-gradient-to-b from-orange-100 to-white">
-      <div className="max-w-2xl mx-auto py-16 px-4 sm:py-24 sm:px-6 lg:max-w-7xl lg:px-8">
-        <h2 className="text-2xl font-extrabold tracking-tight text-gray-900"></h2>
+    <div className="bg-gradient-to-b from-orange-50 to-white">
+
+<main className="max-w-2xl mx-auto pt-16 pb-24 px-4 sm:px-6 lg:max-w-7xl lg:px-8">
+      <h1 className="text-3xl font-extrabold tracking-tight text-gray-900 sm:text-4xl">Products</h1>
 
         <div className="mt-6 grid grid-cols-1 gap-y-10 gap-x-6 sm:grid-cols-2 lg:grid-cols-4 xl:gap-x-8">
           {listableProductList.map((product) => (
@@ -154,7 +155,8 @@ export default function Products() {
             </div>
           ))}
         </div>
-      </div>
+</main>
+
     </div>
   );
 }
