--- app/routes/homepage.tsx
@@ -1,368 +1,0 @@
-import { Dialog, Popover, Tab, Transition } from "@headlessui/react";
-import {
-  MenuIcon,
-  SearchIcon,
-  ShoppingBagIcon,
-  XIcon,
-} from "@heroicons/react/outline";
-import { Link } from "@remix-run/react";
-import serenuplogo from "public/assets/seren-up-logo.png";
-import { Fragment, useState } from "react";
-
-const navigation = {
-  pages: [
-    { name: "Accessories", href: "/accessories" },
-    { name: "Products", href: "/products" },
-  ],
-};
-
-const footerNavigation = {
-  shop: [
-    { name: "Bracelets", href: "/products" },
-    { name: "Accessories", href: "/Accessories" },
-  ],
-  company: [
-    { name: "Who we are", href: "#" },
-    { name: "Terms & Conditions", href: "#" },
-    { name: "Privacy", href: "#" },
-  ],
-  account: [{ name: "Manage Account", href: "#" }],
-  connect: [{ name: "Contact Us", href: "#" }],
-};
-
-export default function Homepage() {
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
-
-  return (
-    <div className="bg-white">
-      {/* Mobile menu */}
-      <Transition.Root show={mobileMenuOpen} as={Fragment}>
-        <Dialog
-          as="div"
-          className="relative z-40 lg:hidden"
-          onClose={setMobileMenuOpen}
-        >
-          <Transition.Child
-            as={Fragment}
-            enter="transition-opacity ease-linear duration-300"
-            enterFrom="opacity-0"
-            enterTo="opacity-100"
-            leave="transition-opacity ease-linear duration-300"
-            leaveFrom="opacity-100"
-            leaveTo="opacity-0"
-          >
-            <div className="fixed inset-0 bg-black bg-opacity-25" />
-          </Transition.Child>
-
-          <div className="fixed inset-0 flex z-40">
-            <Transition.Child
-              as={Fragment}
-              enter="transition ease-in-out duration-300 transform"
-              enterFrom="-translate-x-full"
-              enterTo="translate-x-0"
-              leave="transition ease-in-out duration-300 transform"
-              leaveFrom="translate-x-0"
-              leaveTo="-translate-x-full"
-            >
-              <Dialog.Panel className="relative max-w-xs w-full bg-white shadow-xl pb-12 flex flex-col overflow-y-auto">
-                <div className="px-4 pt-5 pb-2 flex">
-                  <button
-                    type="button"
-                    className="-m-2 p-2 rounded-md inline-flex items-center justify-center text-gray-400"
-                    onClick={() => setMobileMenuOpen(false)}
-                  >
-                    <span className="sr-only">Close menu</span>
-                    <XIcon className="h-6 w-6" aria-hidden="true" />
-                  </button>
-                </div>
-
-                {/* Links */}
-                <Tab.Group as="div" className="mt-2">
-                  <div className="border-b border-gray-200"></div>
-                </Tab.Group>
-
-                <div className="border-t border-gray-200 py-6 px-4 space-y-6">
-                  {navigation.pages.map((page) => (
-                    <div key={page.name} className="flow-root">
-                      <a
-                        href={page.href}
-                        className="-m-2 p-2 block font-medium text-gray-900"
-                      >
-                        {page.name}
-                      </a>
-                    </div>
-                  ))}
-                </div>
-
-                <div className="border-t border-gray-200 py-6 px-4 space-y-6">
-                  <div className="flow-root">
-                    <Link
-                      to="/registration"
-                      className="-m-2 p-2 block font-medium text-gray-900"
-                    >
-                      Sign up
-                    </Link>
-                  </div>
-                  <div className="flow-root">
-                    <Link
-                      to="/login"
-                      className="-m-2 p-2 block font-medium text-gray-900"
-                    >
-                      Sign in
-                    </Link>
-                  </div>
-                </div>
-              </Dialog.Panel>
-            </Transition.Child>
-          </div>
-        </Dialog>
-      </Transition.Root>
-
-      {/* Hero section */}
-      <div className="relative bg-orange-700">
-        {/* Decorative image and overlay */}
-        <div aria-hidden="true" className="absolute inset-0 overflow-hidden">
-          <img
-            src="https://tailwindui.com/img/ecommerce-images/home-page-01-hero-full-width.jpg"
-            alt=""
-            className="w-full h-full object-center object-cover"
-          />
-        </div>
-        <div
-          aria-hidden="true"
-          className="absolute inset-0 bg-gradient-to-b from-orange-600 to-orange-100  opacity-50"
-        />
-
-        {/* Navigation */}
-        <header className="relative z-10">
-          <nav aria-label="Top">
-            {/* Top navigation */}
-            <div className="bg-orange-900">
-              <div className="max-w-7xl mx-auto h-10 px-4 flex items-center justify-between sm:px-6 lg:px-8">
-                {/* Currency selector */}
-
-                <div className="flex items-center space-x-6">
-                  <Link
-                    to="/signin"
-                    className="text-sm font-medium text-white hover:text-gray-100"
-                  >
-                    Sign In
-                  </Link>
-                </div>
-              </div>
-            </div>
-
-            {/* Secondary navigation */}
-            <div className="backdrop-blur-md backdrop-filter bg-opacity-10 bg-white">
-              <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-                <div>
-                  <div className="h-16 flex items-center justify-between">
-                    {/* Logo (lg+) */}
-                    <div className="hidden lg:flex-1 lg:flex lg:items-center">
-                      <a href="#">
-                        <span className="sr-only">Workflow</span>
-                        <img className="h-8 w-auto" src={serenuplogo} alt="" />
-                      </a>
-                    </div>
-
-                    <div className="hidden h-full lg:flex">
-                      {/* Flyout menus */}
-                      <Popover.Group className="px-4 bottom-0 inset-x-0">
-                        <div className="h-full flex justify-center space-x-8">
-                          {navigation.pages.map((page) => (
-                            <a
-                              key={page.name}
-                              href={page.href}
-                              className="flex items-center text-sm font-medium text-white"
-                            >
-                              {page.name}
-                            </a>
-                          ))}
-                        </div>
-                      </Popover.Group>
-                    </div>
-
-                    {/* Mobile menu and search (lg-) */}
-                    <div className="flex-1 flex items-center lg:hidden">
-                      <button
-                        type="button"
-                        className="-ml-2 p-2 text-white"
-                        onClick={() => setMobileMenuOpen(true)}
-                      >
-                        <span className="sr-only">Open menu</span>
-                        <MenuIcon className="h-6 w-6" aria-hidden="true" />
-                      </button>
-
-                      {/* Search */}
-                      <a href="#" className="ml-2 p-2 text-white">
-                        <span className="sr-only">Search</span>
-                        <SearchIcon className="w-6 h-6" aria-hidden="true" />
-                      </a>
-                    </div>
-
-                    {/* Logo (lg-) */}
-                    <a href="#" className="lg:hidden">
-                      <span className="sr-only">Workflow</span>
-                      <img src={serenuplogo} alt="" className="h-8 w-auto" />
-                    </a>
-
-                    <div className="flex-1 flex items-center justify-end">
-                      <div className="flex items-center lg:ml-8">
-                        {/* Help */}
-                        {/* Cart */}
-                        <div className="ml-4 flow-root lg:ml-8">
-                          <Link
-                            to={"/orders/cart"}
-                            className="group -m-2 p-2 flex items-center"
-                          >
-                            <ShoppingBagIcon
-                              className="flex-shrink-0 h-6 w-6 text-white"
-                              aria-hidden="true"
-                            />
-                            <span className="ml-2 text-sm font-medium text-white">
-                              0
-                            </span>
-                            <span className="sr-only">
-                              items in cart, view bag
-                            </span>
-                          </Link>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </nav>
-        </header>
-
-        <div className="relative max-w-3xl mx-auto py-32 px-6 flex flex-col items-center text-center sm:py-64 lg:px-0">
-          <h1 className="text-4xl font-extrabold tracking-tight text-white lg:text-6xl">
-            New arrivals are here
-          </h1>
-          <p className="mt-4 text-xl text-white">
-            The new arrivals have, well, newly arrived. Check out the latest
-            options from our summer small-batch release while they're still in
-            stock.
-          </p>
-          <a
-            href="#"
-            className="mt-8 inline-block bg-white border border-transparent rounded-md py-3 px-8 text-base font-medium text-gray-900 hover:bg-gray-100"
-          >
-            Shop New Arrivals
-          </a>
-        </div>
-      </div>
-      <footer aria-labelledby="footer-heading" className="bg-orange-900">
-        <h2 id="footer-heading" className="sr-only">
-          Footer
-        </h2>
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="py-20 xl:grid xl:grid-cols-3 xl:gap-8">
-            <div className="grid grid-cols-2 gap-8 xl:col-span-2">
-              <div className="space-y-12 md:space-y-0 md:grid md:grid-cols-2 md:gap-8">
-                <div>
-                  <h3 className="text-sm font-medium text-white">Shop</h3>
-                  <ul role="list" className="mt-6 space-y-6">
-                    {footerNavigation.shop.map((item) => (
-                      <li key={item.name} className="text-sm">
-                        <a
-                          href={item.href}
-                          className="text-gray-300 hover:text-white"
-                        >
-                          {item.name}
-                        </a>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-                <div>
-                  <h3 className="text-sm font-medium text-white">Company</h3>
-                  <ul role="list" className="mt-6 space-y-6">
-                    {footerNavigation.company.map((item) => (
-                      <li key={item.name} className="text-sm">
-                        <a
-                          href={item.href}
-                          className="text-gray-300 hover:text-white"
-                        >
-                          {item.name}
-                        </a>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              </div>
-              <div className="space-y-12 md:space-y-0 md:grid md:grid-cols-2 md:gap-8">
-                <div>
-                  <h3 className="text-sm font-medium text-white">Account</h3>
-                  <ul role="list" className="mt-6 space-y-6">
-                    {footerNavigation.account.map((item) => (
-                      <li key={item.name} className="text-sm">
-                        <a
-                          href={item.href}
-                          className="text-gray-300 hover:text-white"
-                        >
-                          {item.name}
-                        </a>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-                <div>
-                  <h3 className="text-sm font-medium text-white">Connect</h3>
-                  <ul role="list" className="mt-6 space-y-6">
-                    {footerNavigation.connect.map((item) => (
-                      <li key={item.name} className="text-sm">
-                        <a
-                          href={item.href}
-                          className="text-gray-300 hover:text-white"
-                        >
-                          {item.name}
-                        </a>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              </div>
-            </div>
-            <div className="mt-12 md:mt-16 xl:mt-0">
-              <h3 className="text-sm font-medium text-white">
-                Sign up for our newsletter
-              </h3>
-              <p className="mt-6 text-sm text-gray-300">
-                The latest deals and savings, sent to your inbox weekly.
-              </p>
-              <form className="mt-2 flex sm:max-w-md">
-                <label htmlFor="email-address" className="sr-only">
-                  Email address
-                </label>
-                <input
-                  id="email-address"
-                  type="text"
-                  autoComplete="email"
-                  required
-                  className="appearance-none min-w-0 w-full bg-white border border-white rounded-md shadow-sm py-2 px-4 text-base text-gray-900 placeholder-gray-500 focus:outline-none focus:border-white focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-900 focus:ring-white"
-                />
-                <div className="ml-4 flex-shrink-0">
-                  <button
-                    type="submit"
-                    className="w-full bg-orange-600 border border-transparent rounded-md shadow-sm py-2 px-4 flex items-center justify-center text-base font-medium text-white hover:bg-orange-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-white focus:ring-orange-500"
-                  >
-                    Sign up
-                  </button>
-                </div>
-              </form>
-            </div>
-          </div>
-
-          {/*references to seren-up and Systax */}
-          <div className="border-t border-gray-800 py-10">
-            <p className="text-sm text-gray-400">
-              Copyright &copy; 2022 Seren-Up S.p.A. - Made By Systax
-            </p>
-          </div>
-        </div>
-      </footer>
-    </div>
-  );
-}
