+++ app/routes/signin.tsx
@@ -1,4 +1,4 @@
-import { ActionFunction } from "@remix-run/node";
+import { ActionFunction, redirect } from "@remix-run/node";
 import { Form } from "@remix-run/react";
 import React from "react";
 import { hashPassword } from "~/helpers/crypto.server";
@@ -26,6 +26,41 @@ export const action: ActionFunction = async ({ request }) => {
       password: hashPassword(password),
     });
 
+    const response = await fetch(
+      "http://127.0.0.1:3001/api/v1/ecommerce/login",
+      {
+        method: "post",
+        body,
+      }
+    );
+
+    if (response.ok) {
+      const responseBody = await response.json();
+
+      throw redirect("/homepage");
+    }
+  } else {
+    const email = formBody.get("email");
+    const password = formBody.get("password");
+    const confirmPassword = formBody.get("confirmPassword");
+
+    if (
+      email == null ||
+      password == null ||
+      password !== confirmPassword ||
+      typeof email !== "string" ||
+      typeof password !== "string"
+    ) {
+      return {
+        error: true,
+      };
+    }
+
+    const body = JSON.stringify({
+      email,
+      password: hashPassword(password),
+    });
+
     const response = await fetch("/api/v1/ecommerce/login", {
       method: "post",
       body,
@@ -34,14 +69,12 @@ export const action: ActionFunction = async ({ request }) => {
     if (response.ok) {
       const responseBody = await response.json();
     }
-  } else {
-    
   }
 };
 
 export default function SignIn() {
   const [currentForm, setCurrentForm] = React.useState<"login" | "register">(
-    "login",
+    "login"
   );
 
   if (currentForm === "login") {
@@ -135,7 +168,7 @@ function Login({ setCurrentForm }: LoginProps) {
                   type="submit"
                   name="signin"
                   value="login"
-                  className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
+                  className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-orange-600 hover:bg-orange-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-orange-500"
                 >
                   Login
                 </button>
@@ -170,6 +203,44 @@ function Register({ setCurrentForm }: RegisterProps) {
         <div className="mt-8 sm:mx-auto sm:w-full sm:max-w-md">
           <div className="bg-white py-8 px-4 shadow sm:rounded-lg sm:px-10">
             <Form className="space-y-6" action="#" method="post">
+            <div className="flex flex-row columns-2 justify-between">
+                <div className=" col-span-1 m-1">
+                  <label
+                    htmlFor="text"
+                    className="block text-sm font-medium text-gray-700"
+                  >
+                    Name
+                  </label>
+                  <div className="mt-1 col-span-1">
+                    <input
+                      id="name"
+                      name="name"
+                      type="text"
+                      autoComplete="name"
+                      required
+                      className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
+                    />
+                  </div>
+                </div>
+                <div className="col-span-1 m-1">
+                  <label
+                    htmlFor="text"
+                    className="block text-sm font-medium text-gray-700"
+                  >
+                    Surname
+                  </label>
+                  <div className="mt-1 col-span-1">
+                    <input
+                      id="surname"
+                      name="surname"
+                      type="text"
+                      autoComplete="surname"
+                      required
+                      className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
+                    />
+                  </div>
+                </div>
+              </div>
               <div>
                 <label
                   htmlFor="email"
@@ -207,6 +278,25 @@ function Register({ setCurrentForm }: RegisterProps) {
                   />
                 </div>
               </div>
+              <div>
+                <label
+                  htmlFor="password"
+                  className="block text-sm font-medium text-gray-700"
+                >
+                  Confirm Password
+                </label>
+                <div className="mt-1">
+                  <input
+                    id="password"
+                    name="password"
+                    type="password"
+                    autoComplete="current-password"
+                    required
+                    className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
+                  />
+                </div>
+              </div>
+
 
               <div className="flex items-center justify-between">
                 <div className="text-sm">
@@ -232,7 +322,7 @@ function Register({ setCurrentForm }: RegisterProps) {
                   type="submit"
                   name="signin"
                   value="register"
-                  className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
+                  className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-orange-600 hover:bg-orange-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-orange-500"
                 >
                   Register
                 </button>
